// src/app/university-dashboard/data-access/models/student-details-model.ts
import { QuestionnaireResponse } from "../../../shared/data-access/models/questionnaireResponse.model";
import { DocumentData } from "../../../student/models/documentData-model";
import { AdminDocumentData } from "../../../student/models/downloadedAdminDocumentData.model";
import { StudentDocumentData } from "../../../student/models/downloadedStudentDocumentData.model";
import { ApplicationHistory } from "./application-history.model";
import { InvoiceHistory } from "./invoice-history.model ";
import { StudentProfilePicture } from "../../../student/models/photo.model";

export interface UniversityStudentDetails {
  confirmHonors?: string;
  degreeDuration?: number;
  studentId: number;
  name: string;
  surname: string;
  email: string;
  dateOfBirth: string;
  university: string;
  yearOfStudy: number;
  yearOfFunding: number;
  status: string;
  invoiceStatus: string;
  applicationId: number;
  faculty: string;
  accommodation: number;
  tuition: number;
  meals: number;
  other: number;
  otherDescription: string;
  lastStatusChangeDate: string;
  approvedDate: string;
  commencementDate: string;
  amount: number
  applicationDate: string;
  dob: string
  contactNumber: string;
  address: string;
  suburb: string;
  city: string;
  code: string;
  identityDocument: string;
  academicTranscript: string;
  matricCertificate: string;
  financialStatement: string;
  degree: string;
  reason: string;
  motivation: string;
  applicatioMotivation: string;
  profilephoto:string;
  ApplicationHistory?: ApplicationHistory[];
  InvoiceHistory?: InvoiceHistory[];
  fundDistribution?: ApplicationHistory[];
  detailsUpdates?: ApplicationHistory[];
  documentsUpdate?: ApplicationHistory[];
  canRenew?: boolean;
  minAmount:number;
  maxAmount:number;
  AdminDocuments?: DocumentData;
  StudentAverages?: DocumentData;
  QuestionnaireResponses?: QuestionnaireResponse[];
  DownloadedStudentDocuments: StudentDocumentData;
  DownloadedAdminDocuments?: AdminDocumentData;
  StudentProfilePhoto?: StudentProfilePicture;
  applicationGuid: string;
}