// src/app/student/models/documentBlobNames.model.ts
export interface DocumentBlobNames {
    invoiceForTuition?: string;
    invoiceForMeals?: string;
    invoiceForAccommodation?: string;
    invoiceForOther?: string;
    paymentForAccommodation?: string;
    paymentForMeals?: string;
    paymentForOther?: string;
    paymentForTuition?: string;
    contract?: string;
}